+++ src/store/slices/userSettingsSlice.ts
@@ -17,6 +17,7 @@ export interface UserSettingsState {
 	subqueryChaining: string;
 	andOrFilterEnabled: boolean;
 	andOrFilter: string;
+	showAnnouncements: boolean;
 }
 
 const initialState: UserSettingsState = {
@@ -38,6 +39,9 @@ const initialState: UserSettingsState = {
 		? localStorage.getItem(UserSetting.AND_OR_FILTER_ENABLED) == 'true'
 		: false,
 	andOrFilter: localStorage.getItem(UserSetting.AND_OR_FILTER_VALUE) ?? '',
+	showAnnouncements: localStorage.getItem(UserSetting.SHOW_ANNOUNCEMENTS)
+		? localStorage.getItem(UserSetting.SHOW_ANNOUNCEMENTS) == 'true'
+		: true,
 };
 
 export const userSettingsSlice = createSlice({
@@ -59,6 +63,10 @@ export const userSettingsSlice = createSlice({
 				UserSetting.CB_PASSWORD,
 				action.payload.password
 			);
+			//* make sure that the announcements page doesn't show up for new users
+			//* because the condition for it to show is that a username is cached
+			//* and the showAnnouncements value is true (which it is by default)
+			setShowAnnouncements(false);
 		},
 		setCbqlString: (state, action: PayloadAction<string>) => {
 			state.cbqlString = action.payload;
@@ -189,6 +197,14 @@ export const userSettingsSlice = createSlice({
 			state.cbqlString = cbql;
 			localStorage.setItem(UserSetting.CBQL_STRING, cbql);
 		},
+		setShowAnnouncements: (state, action: PayloadAction<boolean>) => {
+			state.showAnnouncements = action.payload;
+
+			localStorage.setItem(
+				UserSetting.SHOW_ANNOUNCEMENTS,
+				action.payload.toString()
+			);
+		},
 	},
 });
 
@@ -202,6 +218,7 @@ export const {
 	resetCbqlStringToCurrentParameters,
 	setAndOrFilterEnabled,
 	setAndOrFilter,
+	setShowAnnouncements,
 } = userSettingsSlice.actions;
 
 export default userSettingsSlice.reducer;